+++ CHANGELOG.rst
@@ -1,10 +1,16 @@
-0.0.0 (2020-02-11)
+0.0.0 (2024-03-26)
 ------------------
 
 * First release on PyPI.
 
-0.0.5 (2020-02-11)
+0.0.6 (2024-04-03)
 ------------------
 
 * Modify readme.rst file
 * Modify Authors.rst
+
+0.0.7 (2024-04-18)
+------------------
+
+* Adding calculation of Domestic Hot Water - DHW NEed according to the ISO 12831-3
+* Adding example of DHW calculation